Clear Mac compiler warnings. Re #966.

+++ Code/Mantid/Framework/API/inc/MantidAPI/CompositeFunctionMW.h
@@ -56,7 +56,7 @@ public:
 
   void setWorkspace(boost::shared_ptr<const Workspace> ws,const std::string& slicing,bool copyData = true);
   /// Set the workspace
-  void setMatrixWorkspace(boost::shared_ptr<const API::MatrixWorkspace> workspace,int spec,int xMin,int xMax);
+  void setMatrixWorkspace(boost::shared_ptr<const API::MatrixWorkspace> workspace, size_t spec, size_t xMin, size_t xMax);
   /// Returns the function's name
   std::string name()const{return "CompositeFunctionMW";}
   /// Writes itself into a string
